Moved browser utils into single file
