+++ apps/web/util/api.ts
@@ -1,3 +1,5 @@
+import App from "@browser/util/app";
+
 interface Data {
   [key: string]: any;
 }
@@ -18,16 +20,23 @@ export class Server {
   static devCDN = "http://localhost:5000/";
   static prodCDN = "https://cdn.remark.surf/";
 
+  static devWSS = "http://localhost:5500/";
+  static prodWSS = "https://wss.remark.surf/";
+
   static get url() {
-    return process.env.NODE_ENV == "development" ? this.devWeb : this.prodWeb;
+    return App.isDev() ? this.devWeb : this.prodWeb;
   }
 
   static get api() {
-    return process.env.NODE_ENV == "development" ? this.devAPI : this.prodAPI;
+    return App.isDev() ? this.devAPI : this.prodAPI;
   }
 
   static get cdn() {
-    return process.env.NODE_ENV == "development" ? this.devCDN : this.prodCDN;
+    return App.isDev() ? this.devCDN : this.prodCDN;
+  }
+
+  static get wss() {
+    return App.isDev() ? this.devWSS : this.prodWSS;
   }
 }
